// SDKs/twilio.com/twilio_preview/1.36.0/typescript/src/sdk/models/shared/previewdeployeddevicesfleet.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class PreviewDeployedDevicesFleet extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=default_deployment_sid" })
  defaultDeploymentSid?: string;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=links" })
  links?: Map<string, any>;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=unique_name" })
  uniqueName?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}